Start moving in nrf-mac code to hophop
