+++ hophop/src/nrfxlib_mac/mod.rs
@@ -1,0 +1,1 @@
+pub mod error;
